In v2, transaction operation can't be "GENESIS" (#46)
If the "version" of a transaction is "2.0" then "operation" can't be "GENESIS", it can only be "CREATE" or "TRANSFER". This change was made by @shahbazn in commit bigchaindb/bigchaindb@e8e02ca . This PR updates the BigchainDB Transactions Spec v2 (BEP-13) accordingly.

This PR also adds a subsection to the section on transaction validation to note that there are sometimes other rules that are specific to the implementation, with a link to that section. That change was made in both versions of the BigchainDB Transactions Spec, i.e. v1 and v2